Re-implemented based on a non-racy design.
Unfortunately, trying to model the task as a duration-based timeout saw
race conditions where the callback to determine whether to start a timer
was called before the callback to add a new track to the player had run.
This caused an issue where the code believed that no track was playing
when in fact a track would be queued mere milliseconds later. When the
video in question was under a minute long, this led to a state where the
bot would leave the channel a minute early.

I've re-implemented the functionality such that it is invoked as part of
the onTrackStart and onTrackEnd callbacks of Player. This guarantees
that the idle timer kickoff logic runs _after_ all other Player state
updates have completed, which neatly avoids race/ordering bugs.